[feat] Project Modal에 Suspense, Skeleton 추가 - 데이터 로딩 중 사용자 경험 개선을 위해 Suspense와 스켈레톤 UI를 구현 - 프로젝트 모달의 콘텐츠가 렌더링되기 전에 스켈레톤 로딩 상태를 보여주도록 설정
