+++ src/app/@modal/(.)[slug]/page.tsx
@@ -1,6 +1,4 @@
 import { ProjectModal } from "@/components/ui/ProjectModal";
-import { getProjectBySlug } from "@/lib/projects";
-import { notFound } from "next/navigation";
 
 interface ModalPageProps {
   params: { slug: string };
@@ -10,11 +8,5 @@ export default async function InterceptedProjectModal({
   params,
 }: ModalPageProps) {
   const { slug } = await params;
-  const project = await getProjectBySlug(slug);
-
-  if (!project) {
-    notFound();
-  }
-
-  return <ProjectModal project={project} />;
+  return <ProjectModal slug={slug} />;
 }
