+++ src/app/[slug]/page.tsx
@@ -1,6 +1,4 @@
 import { ProjectModal } from "@/components/ui/ProjectModal";
-import { getProjectBySlug } from "@/lib/projects";
-import { notFound } from "next/navigation";
 
 interface ProjectPageProps {
   params: { slug: string };
@@ -8,16 +6,11 @@ interface ProjectPageProps {
 
 export default async function ProjectPage({ params }: ProjectPageProps) {
   const { slug } = await params;
-  const project = await getProjectBySlug(slug);
-
-  if (!project) {
-    notFound();
-  }
 
   return (
     <main className="min-h-screen bg-white dark:bg-gray-900">
       <div className="container mx-auto px-4 py-8">
-        <ProjectModal project={project} />
+        <ProjectModal slug={slug} />
       </div>
     </main>
   );
